Helm: configure the agent image (#919)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

This PR enhances the Helm chart such that it automatically configures
the operator as to which agent image will be used. The image that is
used shall always be the operator's own image, as opposed to having a
separate entry in values.yaml.

Here's the values that control which image is used as both the operator
image and as the agent image:

```yaml
image:
  # -- The image registry to pull from
  registry: docker.io
  # -- The image repository to pull from
  repository: pulumi/pulumi-kubernetes-operator
  # -- The image pull policy
  pullPolicy: IfNotPresent
  # -- The image tag to pull, default: `Chart.appVersion`
  tag: ""
```

This is an improvement over the status quo that the agent image is not
configurable at all, which is a problem when a non-default operator
image is used.

Important Note: this enhancement is usable only with Helm-based
deployments. The alternative ("quickstart") deployment is not
configurable anyway.

### Walkthrough

Install the operator with a non-default image tag of `v2.1.0-devel` and
with an image pull policy of `Never`:

```
helm install pulumi-kubernetes-operator ./deploy/helm/pulumi-operator  --namespace pulumi-kubernetes-operator \
  --set image.tag=v2.1.0-devel --set image.pullPolicy=Never
```

Deploy a Workspace pod and observe that the init containers are using
the configured image and policy:
```yaml
  apiVersion: v1
  kind: Pod
  spec:
    initContainers:
    - name: bootstrap
      image: docker.io/pulumi/pulumi-kubernetes-operator:v2.1.0-devel
      imagePullPolicy: Never
      name: bootstrap
```

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #848 (thanks @project0)