Stabilize the ordering of initContainers (#923)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

This PR is a breaking API change to ensure that a workspace's built-in
init containers (bootstrap, fetch) are declared before any extra init
containers found in the `podTemplate`.

Implementation-wise, the desired ordering is achieved with the
`$setElementOrder` directive, as documented here:

https://github.com/kubernetes/design-proposals-archive/blob/main/cli/preserve-order-in-strategic-merge-patch.md

### Migration Notes
If your application uses the `podTemplate` to inject init container(s),
be aware that they'll run after the `bootstrap` and `fetch` containers.
This means that the `/share` volume will be present, and your program
code will have been downloaded, before your init container runs.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #774