Add an `updateTemplate` to the `Stack` spec (#925)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

This PR adds a new field to the `Stack` spec, an `updateTemplate` for
customizing the `Update` objects that are made by the stack controller.
Works similarly to the `workspaceTemplate` field.

The ostensible reason is to allow the retention TTL to be customized:

```yaml
apiVersion: pulumi.com/v1
kind: Stack
metadata:
  name: random-yaml-program
  namespace: default
spec:
  workspaceTemplate:
    spec:
      image: pulumi/pulumi:3.147.0-nonroot
  updateTemplate:
    metadata:
      labels:
        app.kubernetes.io/name: "myapp"
    spec:
      ttlAfterCompleted: 42m
```

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #859