implement dynamic environment variables (#976)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

Implements the suggested solution to automatically read an environment
file (`/share/.env`) into the pulumi container, to be able to
dynamically set environment variables affecting the Pulumi CLI and your
program and to configure providers. For example, cloud hosting providers
set certain environment variables onto the pod that would be the basis
for other variables needed by the program. This extension point makes it
possible.

Unlike variables set thru the Stack spec, these variables are set by
init containers by writing to an _environment file_ that is aliased by
the `PULUMI_ENV` variable. Init containers may write to the environment
file at will to set variables for the Pulumi CLI. This solution is like
[the familiar GITHUB_ENV environment
file](https://docs.github.com/en/actions/how-tos/write-workflows/choose-what-workflows-do/use-variables#passing-values-between-steps-and-jobs-in-a-workflow).

 For example:
```yaml
initContainers:
  - name: extra
    image: busybox
    command:
      - sh
      - -c
      - |
          echo 'PULUMI_CONFIG_PASSPHRASE=test' >> $PULUMI_ENV
```

### Testing

A new e2e test was developed that deploys a stack based on a filesystem
backend, and sets the `PULUMI_CONFIG_PASSPHRASE` variable dynamically.
The stack update is successful only if the dynamic variable is passed as
expected to the CLI.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #937