Preview mode for Stack API (#1012)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->
Proposal is to add a new “preview-only” flag that forces a given Stack
object to **run in preview mode only**. This is NOT
update-with-a-preview.

The basic idea is for the user to **create multiple Stack objects for
the same underlying Pulumi stack**, with all-but-one operating in a
dry-run fashion to illustrate a “what-if” configuration.

Such an arrangement would allow the user to **run previews of a stack
against the same or a different branch/tag**, and obtain full status
such as the preview link, standard output, and program outputs*. The
user may even vary the stack configuration of a given preview.

Another, simpler way to use this feature would be to “toggle” the
preview flag on and off on a given Stack object, to **roll out a change
in a tick-tock fashion** with external verification between turns. The
Stack API has strong support for `metadata.generation` to make sense of
the status (as each spec change would effect a new generation).

The Stack API remains otherwise unchanged. In preview mode, the concept
of “in-sync” (as reflected by the `Ready` condition) indicates that the
preview was successful.

### Testing

1. new e2e test - examples/preview-demo
2. new unit test cases for "preview" flag

### Example

```yaml
apiVersion: pulumi.com/v1
kind: Stack
metadata:
 name: random-yaml
spec:
 stack: eron-pulumi-corp/random-yaml
 projectRepo: https://github.com/pulumi/examples
 branch: v1.0
status:
 permalink: https://app.pulumi.com/eron-pulumi-corp/random/random-yaml-git/updates/1
 conditions:
 - lastTransitionTime: "2025-09-18T23:33:04Z"
   message: the stack has been processed and is up to date
   reason: ProcessingCompleted
   status: "True"
   type: Ready
---
apiVersion: pulumi.com/v1
kind: Stack
metadata:
 name: random-yaml-preview-v2.0
spec:
 stack: eron-pulumi-corp/random-yaml
 projectRepo: https://github.com/pulumi/examples
 branch: v2.0
 preview: true
status:
 permalink: https://app.pulumi.com/eron-pulumi-corp/random/random-yaml-git/previews/fc0c65fd-0609-4c57-8170-ac3edbc89d4b
 conditions:
 - lastTransitionTime: "2025-09-18T23:33:04Z"
   message: the stack has been previewed
   reason: ProcessingCompleted
   status: "True"
   type: Ready
```

The series of updates and previews, with updated messages, resembles:


<img width="502" height="947" alt="image"
src="https://github.com/user-attachments/assets/4e8bffb7-ffc2-4bd7-a8bc-014dcdf2e7f8"
/>


### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #16